implement fork/revised/2

+++ sys/memory_helpers.c
@@ -321,7 +321,7 @@ void _do_mark_COW(uint64_t* table_p, uint64_t* table_c, int pid, int lvl){
 		if(table_p[i]== 0)continue;
 		mem_page* next_lvl_page = get_free_page();
 		vma_register_page(next_lvl_page, pid);
-		table_c[i] = ((uint64_t)next_lvl_page->base|PRESENT|READ_WRITE|USER_ACCESSIBLE);
+		table_c[i] = (next_lvl_page->base|PRESENT|READ_WRITE|USER_ACCESSIBLE);
 		_do_mark_COW(
 			(uint64_t*)((((uint64_t)(table_p))|0xffffff8000000000| (i<<3))<<9),
 			(uint64_t*)((((uint64_t)(table_c))|0xffffff8000000000| (i<<3))<<9),
